+++ app/containers/DownloadButtons.js
@@ -12,6 +12,7 @@ const mapStateToProps = (state) => {
         selectedStartDate: state.selectedSearch.dates.selected.start,
         selectedEndDate: state.selectedSearch.dates.selected.end,
         selectedLocation: state.selectedSearch.locations.selected,
+        selectedFilters: state.searchFilters.selected,
         api: state.backends.selected,
         drawShapeCoordinates: state.sensors.shape_coordinates,
         availableSensors: state.sensors.available_sensors